qca-ssdk: Enable the cpu code exception control
When the decapsulation occurs and exception is generated, there are
two excepton modes.

Exception mode0: whole packet with outter and inner packet is
redirected to CPU with only first descriptor, the outter header is
in the fisrt descriptor.

Exception mode1: only inner packet is redirected to CPU with both
first and second descriptor, and first is inner header, the second
is outer header.

Enable the EXCEPTION_FMT_CTRL_EN of all CPU code by default.
the exception mode is controlled by fal_tunnel_exp_decap_set.

Change-Id: Ie7df6542b44608701967349fb1265963c7f3c7aa
Signed-off-by: Luo Jie <quic_luoj@quicinc.com>