ssdk_clk: Remove MP_PHY clocks resets and init logic
While commit 81d3888 disables the init function for MP_PHY as there's a
native linux driver for the IPQ5018 PHY, the SSDK has the PHY's clock
and reset ops mixed in functions that cover different components.
In addition, the enablement of the LDO controller is part of and taken
care of by the IPQ5018 PHY driver itself.
So let's add preprocessor directives to disable compilation and avoid
unpredictable clock and reset behavior.

Signed-off-by: George Moussalem <george.moussalem@outlook.com>