MP: fix uniphy reset in phy to phy link scenario
Uniphy supports fixed link and phy to phy link architectures.
Prior to below patch issued by QCA to fix the uniphy reset in the GCC,
the reset bit - BIT(0) - triggered an AHB reset which worked well for
fixed link scenarios. After the patch was applied, it was set to a
bitmask, triggering resets for the uniphy SYS, RX, and TX clocks which
are needed in phy to phy link setups to support shifting between link
speeds (1G & 2.5G for MP). This broke fixed links, however. So let's
trigger the right reset based on the link architecture.

Link: https://git.codelinaro.org/clo/qsdk/oss/kernel/linux-ipq-5.4/-/commit/036bdc62aca561e8ff94a29c447fc400de2b7304

Signed-off-by: George Moussalem <george.moussalem@outlook.com>